Rewrote DepthFilter
Depthfilter was a good idea, but sadly Source-limits ruin the effect. Source’s depth is limited to ~192 units and there is no way of increasing it or make a cut-off point. Additionally, some GPU’s / Branches don’t like textures using the depth-texture.

I’ve recreated it to use a sphere-stencil + RT + magic to render it in the distance. There is sadly no smooth transition, but the effect should be good for most cases and will help selling the effect, when players are moving around fast.